Update 03-writing-your-first-liferay-application.markdown
It could use some formatting help on the items I added, but the new code snippet provides the same consistency of "here is what you should have" as previous sections.  For me, going line by line through the instructions, I also found that I was missing the necessary step for the UI taglib import, even if we have done something like it earlier on.