Update 03-configuring-the-patching-tool.markdown
Removed step 3 and 4 as it's not related to the patching tool setup. It's only necessary when the patching tool is moved to a different directory, however that should not be the scope of this document. Also fixed the example patching.tool.home value to a more general one instead of my home folder. :)