Added headers for missing content
This doc page leaves a lot of unanswered questions. This commit is not an update to the docs so much as a request for info that is missing. Please update the doc so that it answers the questions posed in the parens(). I couldn't figure out if there was a way to comment on the file without doing a commit.

Thanks, guys!