Discontinue use of $TOMCAT_HOME in favor of $CATALINA_*
Change directory names from $TOMCAT_HOME to $CATALINA_HOME and $CATALINA_BASE to be synchronized with tomcat's documentation and naming conventions. Great documentation for the difference of both variables (well, mostly they're refering to the same directory anyway) is in tomcat's RUNNING.txt, which is present in every tomcat installation.