Update 01-installing-blade-cli.markdown
Hey guys this is a much simplified way of installing blade cli, its just a one-liner (no more downloading) for those that have JPM installed.  Now if there is a problem with proxies etc, there is the "manual" way of downloading the CLI jar and installing it from disk, but I don't think we need to have that in the official docs.