Package name correction and warning
The packages of Store implementations have changed from 6.2 to 7.0 - this is easy to miss when bringing the properties forward and the error messages one is running in aren't really helpful. With an explicit warning here, hopefully less people will run into this problem. Plus, the documentation contained an old package name, which doesn't help as well...