LRDOCS-3235 Just added some extra notes
I am not sure if you want to add this information. I don't know whether
this information is already present elsewhere. I believe it is important
no note that the standar tooling out there can also be used, so devs out
there don't feel constrained to having to learn all our stack when they
might not want to or might not see the benefit in it.

Maybe we can create example of that too?

Just let me know your thoughts.