Expanded the "app-server.properties" and descriptions
Expanded the "app-server.properties" configuration example to better reflect what the file looks like after manual input. I also expanded the description to inform the user what each line is looking for.

This will allow the upgrade tool to run without asking the user to input the data due to missing information. Additionally, to my current understanding, those lines are usable as is in any DXP upgrade as long as the user's tomcat is tomcat-8.0.32.