Removed superfluous 'it' and confusing parenthesis
The 'it' placed after declarations on line 59 is an obvious typo. I also removed two parenthetical notes "(in some cases" and "(in some others)". I don't believe these parenthetical notes add any context or clarification to the statement. I assume the author was trying to distinguish between or and xor (opting for xor) which might be wrong anyway. If I remember correctly there were instances in 6.2 where the same configuration can be made in portlet.xml as in portal.properties (disabling csrf token for example). Either way if it is decided that xor is the correct logical structure, and that using the colloquial or isn't satisfactory, I still think there would be a clearer way to distinguish it.