Token Use (Access) and Client Cred. and Res Own
Client Cred. and Res Own release an access token without prompting for credentials.

The chapter "Token Use" talk about access token. In the last paragraph it's explained using the words "authorization code" but IMHO it's more appropriate the use of access token (if I'm not wrong the /token endpoint respond always with an access or refresh token).