Added note about osgi.web.symbolicname value
The relation between osgi.web.symbolicname and Bundle-SymbolicName is not straightforward, especially for the beginners who may lose hours trying to figure out why their Collection is not appearing in the editor, while the bundle is properly activated and there are no errors in the console. I think it should be clearly stated they must be the same.